refactor : SOLID-token-interface
