+++ src/test/java/com/sparta/uglymarket/dto/AdminLoginRequestTest.java
@@ -1,17 +1,17 @@
-package com.sparta.uglymarket.dto;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-public class AdminLoginRequestTest {
-
-    @Test
-    void testDefaultConstructorAndGetters() {
-        // given
-        AdminLoginRequest request = new AdminLoginRequest();
-
-        // then
-        assertNull(request.getPhoneNumber());
-        assertNull(request.getPassword());
-    }
-}
+//package com.sparta.uglymarket.dto;
+//
+//import org.junit.jupiter.api.Test;
+//import static org.junit.jupiter.api.Assertions.*;
+//
+//public class AdminLoginRequestTest {
+//
+//    @Test
+//    void testDefaultConstructorAndGetters() {
+//        // given
+//        AdminLoginRequest request = new AdminLoginRequest();
+//
+//        // then
+//        assertNull(request.getPhoneNumber());
+//        assertNull(request.getPassword());
+//    }
+//}
