+++ src/test/java/com/sparta/uglymarket/dto/AdminLoginResponseTest.java
@@ -1,22 +1,22 @@
-package com.sparta.uglymarket.dto;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-public class AdminLoginResponseTest {
-
-    @Test
-    void testAdminLoginResponseConstructorAndGetter() {
-        // given
-        String expectedMessage = "Login successful";
-
-        // when
-        AdminLoginResponse response = new AdminLoginResponse(expectedMessage);
-
-        // then
-        assertNotNull(response);
-        assertEquals(expectedMessage, response.getMessage());
-    }
-}
+//package com.sparta.uglymarket.dto;
+//
+//import org.junit.jupiter.api.Test;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertNotNull;
+//
+//public class AdminLoginResponseTest {
+//
+//    @Test
+//    void testAdminLoginResponseConstructorAndGetter() {
+//        // given
+//        String expectedMessage = "Login successful";
+//
+//        // when
+//        AdminLoginResponse response = new AdminLoginResponse(expectedMessage);
+//
+//        // then
+//        assertNotNull(response);
+//        assertEquals(expectedMessage, response.getMessage());
+//    }
+//}
