+++ src/test/java/com/sparta/uglymarket/dto/AdminRegisterResponseTest.java
@@ -1,24 +1,24 @@
-package com.sparta.uglymarket.dto;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-public class AdminRegisterResponseTest {
-
-    @Test
-    void testAdminRegisterResponseConstructorAndGetters() {
-        // given
-        String expectedMessage = "Registration successful";
-        String expectedRole = "ADMIN";
-
-        // when
-        AdminRegisterResponse response = new AdminRegisterResponse(expectedMessage, expectedRole);
-
-        // then
-        assertNotNull(response);
-        assertEquals(expectedMessage, response.getMessage());
-        assertEquals(expectedRole, response.getRole());
-    }
-}
+//package com.sparta.uglymarket.dto;
+//
+//import org.junit.jupiter.api.Test;
+//
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertNotNull;
+//
+//public class AdminRegisterResponseTest {
+//
+//    @Test
+//    void testAdminRegisterResponseConstructorAndGetters() {
+//        // given
+//        String expectedMessage = "Registration successful";
+//        String expectedRole = "ADMIN";
+//
+//        // when
+//        AdminRegisterResponse response = new AdminRegisterResponse(expectedMessage, expectedRole);
+//
+//        // then
+//        assertNotNull(response);
+//        assertEquals(expectedMessage, response.getMessage());
+//        assertEquals(expectedRole, response.getRole());
+//    }
+//}
