+++ src/test/java/com/sparta/uglymarket/dto/TokenResponseTest.java
@@ -1,23 +1,23 @@
-package com.sparta.uglymarket.dto;
-
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertNotNull;
-
-public class TokenResponseTest {
-
-    @Test
-    void testParameterizedConstructorAndGetters() {
-        // given
-        String expectedAccessToken = "access-token-example";
-        String expectedRefreshToken = "refresh-token-example";
-
-        // when
-        TokenResponse response = new TokenResponse(expectedAccessToken, expectedRefreshToken);
-
-        // then
-        assertNotNull(response);
-        assertEquals(expectedAccessToken, response.getAccessToken());
-        assertEquals(expectedRefreshToken, response.getRefreshToken());
-    }
-}
+//package com.sparta.uglymarket.dto;
+//
+//import org.junit.jupiter.api.Test;
+//import static org.junit.jupiter.api.Assertions.assertEquals;
+//import static org.junit.jupiter.api.Assertions.assertNotNull;
+//
+//public class TokenResponseTest {
+//
+//    @Test
+//    void testParameterizedConstructorAndGetters() {
+//        // given
+//        String expectedAccessToken = "access-token-example";
+//        String expectedRefreshToken = "refresh-token-example";
+//
+//        // when
+//        TokenResponse response = new TokenResponse(expectedAccessToken, expectedRefreshToken);
+//
+//        // then
+//        assertNotNull(response);
+//        assertEquals(expectedAccessToken, response.getAccessToken());
+//        assertEquals(expectedRefreshToken, response.getRefreshToken());
+//    }
+//}
